feat: absolute imports for tests
