+++ packages/database/package.json
@@ -43,7 +43,7 @@
   "scripts": {
     "ci:lint": "eslint .",
     "ci:build": "yarn ci:lint && yarn build && yarn build:test",
-    "ci:restore": "NODE_ENV=test ts-node ./tests/server/restore.ts",
+    "ci:restore": "NODE_ENV=test ts-node ./tests/restore-server.ts",
     "ci:start": "NODE_ENV=test node dist/server",
     "ci:dev": "yarn build && yarn ci:start",
     "ci:test:database": "yarn test:database",
