+++ packages/database/tests/configs/jest.config.ts
@@ -2,6 +2,18 @@ import { Config } from '@jest/types'
 
 export const all: Config.InitialOptionsWithRootDir = {
   rootDir: '../..',
+  moduleNameMapper: {
+    '@entity': '<rootDir>/src/entity',
+    '@repository': '<rootDir>/src/repository',
+    '@utils': '<rootDir>/src/utils',
+    '@config': '<rootDir>/src/config',
+    '@sql': '<rootDir>/src/sql',
+    '@environment': '<rootDir>/tests/environment',
+    '@init': '<rootDir>/tests/init',
+    '@src/(.*)': '<rootDir>/src/$1',
+    '@mtypes/(.*)': '<rootDir>/types/$1',
+    '@tests/(.*)': '<rootDir>/tests/$1',
+  },
   // show console logs
   silent: true,
   // shows results of each test function within each test file
@@ -12,8 +24,8 @@ export const all: Config.InitialOptionsWithRootDir = {
   clearMocks: true,
   // actual stuff
   preset: 'ts-jest',
-  testMatch: ['**/tests/**/*.test.ts'],
-  testSequencer: './tests/configs/sequencer.js',
+  testMatch: ['<rootDir>/tests/**/*.test.ts'],
+  testSequencer: '<rootDir>/tests/configs/sequencer.js',
 }
 
 /**
