+++ packages/database/tests/degree/insert-modules.test.ts
@@ -1,7 +1,7 @@
-import { container, getSource } from '../../src/data-source'
-import { init } from '../init'
-import { setup, teardown, Repo, t } from '../environment'
-import { flatten, oneUp } from '../../src/utils'
+import { container, getSource } from '@src/data-source'
+import { init } from "@tests/init"
+import { setup, teardown, Repo, t } from '@environment'
+import { flatten, oneUp } from '@utils'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
