+++ packages/database/tests/graph/initialize/no-pull-all.test.ts
@@ -1,7 +1,7 @@
-import { getSource } from '../../../src/data-source'
-import { oneUp } from '../../../src/utils'
-import { setup, teardown, Repo, t } from '../../environment'
-import { init } from '../../init'
+import { getSource } from '@src/data-source'
+import { oneUp } from '@utils'
+import { setup, teardown, Repo, t } from '@environment'
+import { init } from "@tests/init"
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
