+++ packages/database/tests/graph/initialize/pull-all.test.ts
@@ -1,8 +1,8 @@
-import { flatten, oneUp } from '../../../src/utils'
-import { container, getSource } from '../../../src/data-source'
-import { Graph } from '../../../src/entity'
-import { setup, teardown, Repo, t } from '../../environment'
-import { init } from '../../init'
+import { flatten, oneUp } from '@utils'
+import { container, getSource } from '@src/data-source'
+import { Graph } from '@entity'
+import { setup, teardown, Repo, t } from '@environment'
+import { init } from "@tests/init"
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
