+++ packages/database/tests/graph/suggest/from-one.test.ts
@@ -1,9 +1,9 @@
-import { container, endpoint, getSource } from '../../../src/data-source'
-import { Module } from '../../../src/entity'
-import { setup, teardown, Repo, t } from '../../environment'
-import { InitProps } from '../../../types/init-props'
-import { init } from '../../init'
-import { oneUp } from '../../../src/utils'
+import { container, endpoint, getSource } from '@src/data-source'
+import { Module } from '@entity'
+import { setup, teardown, Repo, t } from '@environment'
+import { InitProps } from '@mtypes/init-props'
+import { init } from "@tests/init"
+import { oneUp } from '@utils'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
