+++ packages/database/tests/module/fetch.test.ts
@@ -1,7 +1,7 @@
-import { container, getSource } from '../../src/data-source'
-import { Module } from '../../src/entity'
-import { oneUp } from '../../src/utils'
-import {  setup, teardown, Repo } from '../environment'
+import { container, getSource } from '@src/data-source'
+import { Module } from '@entity'
+import { oneUp } from '@utils'
+import {  setup, teardown, Repo } from '@environment'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
