+++ packages/database/tests/restore-server.ts
@@ -1,0 +1,4 @@
+import { config } from '../src/config'
+import { sql } from '../src/sql'
+
+sql.restoreFromFile(config.database, config.restoreSource)
