+++ packages/database/tests/server/graph/delete.test.ts
@@ -1,6 +1,6 @@
 import { AxiosError } from 'axios'
-import { init } from '../../init'
-import { Create, Delete, server, t } from '../environment'
+import { init } from "@tests/init"
+import { Create, Delete, server, t } from '@environment'
 
 beforeAll(() =>
   Promise.all([Create.User(init.user2), Create.Degree(init.degree1)])
