+++ packages/database/tests/server/graph/list.test.ts
@@ -1,6 +1,6 @@
-import { Graph } from '../../../src'
-import { init } from '../../init'
-import { Create, Delete, server, t } from '../environment'
+import { Graph } from '@entity'
+import { init } from "@tests/init"
+import { Create, Delete, server, t } from '@environment'
 
 beforeAll(() =>
   Promise.all([Create.User(init.user2), Create.Degree(init.degree1)])
