+++ packages/database/tests/server/graph/toggle.test.ts
@@ -1,7 +1,7 @@
 import { AxiosError } from 'axios'
-import { ResponseProps } from '../../../types/api-response'
-import { init } from '../../init'
-import { Create, Delete, server } from '../environment'
+import { ResponseProps } from '@mtypes/api-response'
+import { init } from "@tests/init"
+import { Create, Delete, server } from '@environment'
 
 const t: Partial<{
   graph: ResponseProps['Graph']
