+++ packages/database/tests/server/module-condensed.test.ts
@@ -1,4 +1,4 @@
-import { server } from './environment'
+import { server } from '@environment'
 
 /**
  * try a specific full module search
