--- packages/database/tests/server/restore.ts
@@ -1,4 +1,0 @@
-import { config } from '../../src/config'
-import { sql } from '../../src/sql'
-
-sql.restoreFromFile(config.database, config.restoreSource)
