+++ packages/database/tests/server/user/create.test.ts
@@ -1,7 +1,7 @@
 import { AxiosError } from 'axios'
-import { User } from '../../../src/entity'
-import { Delete, server, t } from '../environment'
-import { init } from '../../init'
+import { User } from '@entity'
+import { Delete, server, t } from '@environment'
+import { init } from "@tests/init"
 
 afterAll(() => Delete.User(t.userId))
