+++ packages/database/tests/server/user/list.test.ts
@@ -1,7 +1,7 @@
-import { User } from '../../../src/entity'
-import { Create, Delete, server, t } from '../environment'
-import { init } from '../../init'
-import { toBeUserResponse } from '../expect-extend'
+import { User } from '@entity'
+import { Create, Delete, server, t } from '@environment'
+import { init } from "@tests/init"
+import { toBeUserResponse } from '@tests/custom-expect'
 
 beforeAll(async () => {
   expect.extend({ toBeUserResponse })
