+++ packages/database/tests/user/getPostReqs/no-optional-params.test.ts
@@ -1,9 +1,9 @@
-import { container, getSource } from '../../../src/data-source'
-import { Module } from '../../../src/entity'
-import { InitProps } from '../../../types/init-props'
-import { init } from '../../init'
-import { setup, teardown, t, Repo } from '../../environment'
-import { flatten, oneUp } from '../../../src/utils'
+import { container, getSource } from '@src/data-source'
+import { Module } from '@entity'
+import { InitProps } from '@mtypes/init-props'
+import { init } from "@tests/init"
+import { setup, teardown, t, Repo } from '@environment'
+import { flatten, oneUp } from '@utils'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
