+++ packages/database/tests/user/getUnlockedModules.test.ts
@@ -1,8 +1,8 @@
-import { container, getSource } from '../../src/data-source'
-import { InitProps } from '../../types/init-props'
-import { init } from '../init'
-import { setup, teardown, Repo, t } from '../environment'
-import { flatten, oneUp } from '../../src/utils'
+import { container, getSource } from '@src/data-source'
+import { InitProps } from '@mtypes/init-props'
+import { init } from "@tests/init"
+import { setup, teardown, Repo, t } from '@environment'
+import { flatten, oneUp } from '@utils'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
