+++ packages/database/tests/user/initialize.test.ts
@@ -1,8 +1,8 @@
-import { container, getSource } from '../../src/data-source'
-import { User } from '../../src/entity'
-import { setup, teardown, Repo, t } from '../environment'
-import { init } from '../init'
-import { oneUp } from '../../src/utils'
+import { container, getSource } from '@src/data-source'
+import { User } from '@entity'
+import { setup, teardown, Repo, t } from '@environment'
+import { init } from "@tests/init"
+import { oneUp } from '@utils'
 
 const dbName = oneUp(__filename)
 const db = getSource(dbName)
