+++ packages/database/tests/utils/string/getModuleLevel.test.ts
@@ -1,4 +1,4 @@
-import { getModuleLevel } from '../../../src/utils'
+import { getModuleLevel } from '@utils'
 
 describe('String.getModuleLevel', () => {
   it('Correctly extracts the 4-digit code', () => {
